lesson2

+++ Python_project/lesson_2/ex_5.py
@@ -1,5 +1,5 @@
 my_list = [54, 37, 24, 15, 8]
-n = float(input("Введите натуральное число (целое число >0): "))
+n = int(input("Введите натуральное число (целое число >0): "))
 if n > 0:
     my_list.append(n)
     my_list.sort(reverse=True)
